+++ tests/unit/datazip_test.py
@@ -1,4 +1,8 @@
 """Core :class:`.DataZip` tests."""
+from datetime import datetime
+from pathlib import Path
+from typing import NamedTuple
+
 import numpy as np
 import pandas as pd
 import pytest
@@ -23,102 +27,39 @@ def test_dfs_to_from_zip(temp_dir):
         ),
         "b": pd.Series([1, 2, 3, 4]),
     }
-    try:
-        DataZip.dfs_to_zip(
-            temp_dir / "test_dfs_to_from_zip",
-            df_dict,
-        )
-        df_load = DataZip.dfs_from_zip(temp_dir / "test_dfs_to_from_zip")
-        for a, b in zip(df_dict.values(), df_load.values()):
-            assert a.compare(b).empty
-    except Exception as exc:
-        raise AssertionError("Something broke") from exc
+    DataZip.dfs_to_zip(
+        temp_dir / "test_dfs_to_from_zip",
+        df_dict,
+    )
+    df_load = DataZip.dfs_from_zip(temp_dir / "test_dfs_to_from_zip")
+    for a, b in zip(df_dict.values(), df_load.values()):
+        assert a.compare(b).empty
 
 
-def test_datazip(temp_dir):
-    """Test :class:`.DataZip`."""
-    df_dict = {
-        "a": pd.DataFrame(
-            [[0, 1], [2, 3]],
-            columns=pd.MultiIndex.from_tuples([(0, "a"), (1, "b")]),
-        ),
-        "b": pd.Series([1, 2, 3, 4]),
-    }
-    try:
-        with DataZip(temp_dir / "test_datazip.zip", "w") as z:
-            z.writed("a", df_dict["a"])
-            z.writed("b", df_dict["b"])
-            z.writed("c", {1: 3, "3": "fifteen", 5: (0, 1)})
-            z.writed("aa", df_dict["a"].loc[:, (0, "a")])
-            z.writed("aaa", df_dict["a"].loc[:, [(0, "a")]])
-            z.writed("d", "hello world")
-            with pytest.raises(FileExistsError):
-                z.writed("c", {1: 3, "3": "fifteen", 5: (0, 1)})
-            with pytest.raises(FileExistsError):
-                z.writed("b", df_dict["b"])
-    except Exception as exc:
-        raise AssertionError("Something broke") from exc
-    else:
-        with DataZip(temp_dir / "test_datazip.zip", "r") as z1:
-            for n in ("a", "b", "c", "d"):
-                assert n in z1._contents
-            assert "a" in z1._no_pqt_cols
-            assert isinstance(z1.read("a"), pd.DataFrame)
-            assert isinstance(z1.read("b"), pd.Series)
-            assert isinstance(z1.read("aa"), pd.Series)
-            aaa = z1.read("aaa")
-            assert isinstance(aaa, pd.DataFrame)
-            assert isinstance(aaa.columns, pd.MultiIndex)
-            assert isinstance(z1.read("c"), dict)
-            assert isinstance(z1.read("d"), str)
-            assert z1.read("d") == "hello world"
+def test_existing_name(temp_dir):
+    """Test existing name protection."""
+    with DataZip(temp_dir / "test_datazip.zip", "w") as z:
+        z["a"] = "a"
+        with pytest.raises(FileExistsError):
+            z["a"] = 5
 
 
 def test_datazip_contains_len(temp_dir):
     """Test override of ``in``."""
     with DataZip(temp_dir / "test_datazip_contains_len.zip", "w") as z:
-        z.writed("a", pd.Series([1, 2, 3, 4]))
+        z["a"] = pd.Series([1, 2, 3, 4])
     with DataZip(temp_dir / "test_datazip_contains_len.zip", "r") as z1:
         assert "a" in z1
         assert "a.parquet" in z1
         assert len(z) == 1
 
 
-def test_datazip_set_get(temp_dir):
-    """Test item access."""
-    series = pd.Series([1, 2, 3, 4])
-    with DataZip(temp_dir / "test_datazip_set_get.zip", "w") as z:
-        z["a"] = series
-    with DataZip(temp_dir / "test_datazip_set_get.zip", "r") as z1:
-        s1 = z1["a"]
-        pd.testing.assert_series_equal(series, s1, check_names=False)
-
-
-def test_datazip_meta_safety(temp_dir):
+@pytest.mark.parametrize("protected", ["__metadata__", "__attributes__"], ids=idfn)
+def test_datazip_meta_safety(temp_dir, protected):
     """Test :class:`.DataZip`."""
-    try:
-        with DataZip(temp_dir / "test_datazip_meta_safety.zip", "w") as z:
-            with pytest.raises(ValueError):
-                z.writed("__metadata__", {1: 3, "3": "fifteen", 5: (0, 1)})
-            with pytest.raises(ValueError):
-                z.writed("__attributes__", {1: 3, "3": "fifteen", 5: (0, 1)})
-    except Exception as exc:
-        raise AssertionError("Something broke") from exc
-
-
-def test_datazip_numpy(temp_dir):
-    """Test :class:`.DataZip`."""
-    array = np.array([[0.0, 4.1], [3.2, 2.1]])
-    try:
-        with DataZip(temp_dir / "test_datazip_numpy.zip", "w") as z:
-            z.writed("np", array)
-    except Exception as exc:
-        raise AssertionError("Something broke") from exc
-    else:
-        with DataZip(temp_dir / "test_datazip_numpy.zip", "r") as z1:
-            ar = z1.read("np")
-            assert isinstance(ar, np.ndarray)
-            assert np.all(ar == array)
+    with DataZip(temp_dir / f"test_datazip_protection_{protected}.zip", "w") as z:
+        with pytest.raises(ValueError):
+            z[protected] = 3
 
 
 def test_datazip_w(temp_dir):
@@ -146,79 +87,6 @@ def test_datazip_w(temp_dir):
         with pytest.raises(FileExistsError):
             with DataZip(temp_dir / "test_datazip_w.zip", "w") as z3:
                 z3.namelist()
-
-
-def test_timestamp_in_datazip(temp_dir):
-    """Test creating a datazip with a :class:`utils.Timestamp` in it."""
-    dm = pd.Timestamp("2022-10-21")
-    df = pd.DataFrame(
-        [[0, 1], [2, 3]],
-        columns=pd.MultiIndex.from_tuples([(0, "a"), (1, "b")]),
-    )
-
-    with DataZip(temp_dir / "test_timestamp_in_datazip.zip", "w") as z0:
-        z0.writed("df", df)
-        z0.writed("dm", dm)
-    with DataZip(temp_dir / "test_timestamp_in_datazip.zip", "r") as self:
-        dm1 = self.read("dm")
-        assert isinstance(dm1, pd.Timestamp)
-
-
-def test_nested_obj_in_datazip(temp_dir):
-    """Test creating a datazip with a :class:`utils.Timestamp` in it."""
-    dt = {"a": pd.Timestamp("2022-10-21")}
-    df = pd.DataFrame(
-        [[0, 1], [2, 3]],
-        columns=pd.MultiIndex.from_tuples([(0, "a"), (1, "b")]),
-    )
-    ddf = {
-        "a": pd.DataFrame(
-            [[0, 100], [2, 3]],
-            columns=pd.MultiIndex.from_tuples([(0, "a"), (1, "b")]),
-        )
-    }
-
-    with DataZip(temp_dir / "test_nested_obj_in_datazip.zip", "w") as z0:
-        z0.writed("df", df)
-        z0.writed("dt", dt)
-        z0.writed("ddf", ddf)
-    with DataZip(temp_dir / "test_nested_obj_in_datazip.zip", "r") as self:
-        dt1 = self.read("dt")
-        assert isinstance(dt1["a"], pd.Timestamp)
-
-
-def test_nested2_obj_in_datazip(temp_dir):
-    """Test creating a datazip with a :class:`utils.Timestamp` nested in it."""
-    lt = ["a", {"time": pd.Timestamp("2022-10-21")}]
-
-    with DataZip(temp_dir / "test_nested2_obj_in_datazip.zip", "w") as z0:
-        z0.writed("lt", lt)
-        pass
-    with DataZip(temp_dir / "test_nested2_obj_in_datazip.zip", "r") as self:
-        lt1 = self.read("lt")
-        assert isinstance(lt1, list)
-        assert isinstance(lt1[0], str)
-        assert isinstance(lt1[1], dict)
-        assert isinstance(lt1[1]["time"], pd.Timestamp)
-
-
-def test_namedtuple_in_datazip(temp_dir):
-    """Test creating a datazip with a :class:`typing.NamedTuple` in it."""
-    df = pd.DataFrame(
-        [[0, 1], [2, 3]],
-        columns=pd.MultiIndex.from_tuples([(0, "a"), (1, "b")]),
-    )
-    om = ObjMeta("this", "that")
-    with DataZip(temp_dir / "test_namedtuple_in_datazip.zip", "w") as z0:
-        z0.writed("df", df)
-        z0.writed("om", om)
-        z0.writed("tup", (1, 2, (1, 2, 3)))
-        z0.writed("l", [om, om])
-        z0.writed("d", {"a": (1, 2, 3), "b": [32, 45], 3: 4})
-    with DataZip(temp_dir / "test_namedtuple_in_datazip.zip", "r") as self:
-        om1 = self.read("om")
-        self.read("tup")
-        assert isinstance(om1, ObjMeta)
 
 
 def test_datazip_dump_load(temp_dir):
@@ -297,6 +165,32 @@ def test_tuple_in_testklass(temp_dir):
     assert obj1.a == 5
 
 
+def test_namedtuple_fallback(temp_dir):
+    """Test named tuple fallback."""
+
+    class NTF(NamedTuple):
+        a: int
+        b: int
+
+    with DataZip(temp_dir / "test_namedtuple_fallback.zip", "w") as z0:
+        z0["a"] = NTF(a=3, b=3)
+    with DataZip(temp_dir / "test_namedtuple_fallback.zip", "r") as z1:
+        assert z1["a"] == (3, 3)
+
+
+def test_sqlalchemy(temp_dir):
+    """Test sqlalchemy engine."""
+    sqlalchemy = pytest.importorskip("sqlalchemy")
+
+    a = sqlalchemy.create_engine("sqlite:///" + str(Path.home()))
+
+    with DataZip(temp_dir / "test_sqa.zip", "w") as z0:
+        z0["a"] = a
+    with DataZip(temp_dir / "test_sqa.zip", "r") as z1:
+        assert isinstance(z1["a"], type(a))
+        assert z1["a"].url == a.url
+
+
 @pytest.mark.parametrize(
     "key, expected",
     [
@@ -307,13 +201,62 @@ def test_tuple_in_testklass(temp_dir):
         ("string", "this"),
         ("int", 3),
         ("bool", True),
+        ("none", None),
         ("complex", {"d": 2 + 3j}),
+        ("datetime", datetime.now()),
+        ("pdTimestamp", pd.Timestamp(datetime.now())),
+        ("npdatetime64", np.datetime64(datetime.now())),
+        ("namedtuple_nested", [(ObjMeta("this", "that"), 5), 4, (1, 2)]),
+        ("namedtuple", ObjMeta("this", "that")),
+        ("np.array", np.array([[0.0, 4.1], [3.2, 2.1]])),
+        (
+            "df",
+            pd.DataFrame(
+                [[0, 1], [2, 3]],
+                columns=pd.MultiIndex.from_tuples(
+                    [(0, "a"), (1, "b")], names=["l1", "l2"]
+                ),
+            ),
+        ),
+        ("series", pd.Series([1, 2, 3, 4], name="series")),
+        ("TestKlass", TestKlass(a=5, b={"c": (2, 3.5)}, c=5.5)),
+        ("path", Path.home()),
     ],
     ids=idfn,
 )
-def test_special_types(temp_dir, key, expected):
+def test_types(temp_dir, key, expected):
     """Test preservation of tuples."""
-    with DataZip(temp_dir / f"test_special_types_{key}.zip", "w") as z0:
+    with DataZip(temp_dir / f"test_types_{key}.zip", "w") as z0:
         z0[key] = expected
-    with DataZip(temp_dir / f"test_special_types_{key}.zip", "r") as z1:
-        assert z1[key] == expected
+    with DataZip(temp_dir / f"test_types_{key}.zip", "r") as z1:
+        read = z1[key]
+        if not isinstance(read, type(expected)):
+            raise AssertionError(
+                f"test_types for {key} {type(read)} != {type(expected)}"
+            )
+        if isinstance(expected, pd.Series):
+            pd.testing.assert_series_equal(read, expected)
+        if isinstance(expected, pd.DataFrame):
+            pd.testing.assert_frame_equal(read, expected)
+        else:
+            assert np.all(read == expected)
+
+
+@pytest.mark.parametrize("save_old", [True, False], ids=idfn)
+def test_replace(temp_dir, save_old):
+    """Test file-based replace."""
+    file = temp_dir / f"test_replace_{save_old}.zip"
+    with DataZip(file, "w") as z0:
+        z0["a"] = None
+        z0["b"] = {"foo": None, "bar": 3}
+    with DataZip.replace(file, save_old=save_old) as z1:
+        assert "a" in z1
+        z1["c"] = {"this": "that"}
+    if save_old:
+        assert (temp_dir / f"test_replace_{save_old}_old.zip").exists()
+    else:
+        assert not (temp_dir / f"test_replace_{save_old}_old.zip").exists()
+    with DataZip(file, "r") as z2:
+        assert z2["a"] is None
+        assert z2["b"] == {"foo": None, "bar": 3}
+        assert z2["c"] == {"this": "that"}